# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API fast-learning",
        "description": "Bienvenue sur la version 1 de la documentation API de fast-learning.<br><br><u>Condition préalable :</u><br><li>Avoir un compte utilisateur</li><br><li>Une clé d'application <b>apiKey</b> vous sera donné lorsque que vous vous connectez. Utilisé là pour faire des requêtes</li><br>",
        "contact": {
            "name": ": TODO AFTER THE SERVER IS ONLINE",
            "email": "elian.yutz@hotmail.fr"
        },
        "license": {
            "name": "TODO AFTER THE SERVER IS ONLINE",
            "url": "https:google.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api",
            "description": "Serveur API Rest"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Permet de se créer un compte utilisateur",
                "description": "Permet de se créer un compte utilisateur. Il faut renseigner l'adresse mail et le username qui doit être des identifiants uniques, le mot de passe, le prénom et le nom",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Adresse mail. A renseigner afin d'obtenir son token utilisateur (api_token (JWT))",
                                        "type": "string",
                                        "example": "johnnn@example.com"
                                    },
                                    "username": {
                                        "description": "l'username",
                                        "type": "string",
                                        "example": "Jojon"
                                    },
                                    "first_name": {
                                        "description": "Prénom de l'utilisateur",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "description": "Nom de l'utilisateur",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "password": {
                                        "description": "Mot de passe.",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Permet de se connecter à l’API fast-learning",
                "description": "Permet de se connecter à l’API.  Il faut renseigner une adresse mail/un mot de passe",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Adresse mail",
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Permet de se déconnecter de l’API fast-learning",
                "description": "Permet de se déconnecter de l’API.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}